feat(): add support for standalone pages this introduces a new standalone flag which can generate a standalone page for routing
