+++ packages/schematics/page/files/__name@dasherize@if-flat__/__name@dasherize__.page.spec.ts
@@ -1,18 +1,11 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { <%= classify(name) %>Page } from './<%= dasherize(name) %>.page';
 
 describe('<%= classify(name) %>Page', () => {
   let component: <%= classify(name) %>Page;
   let fixture: ComponentFixture<<%= classify(name) %>Page>;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ <%= classify(name) %>Page ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
+  beforeEach(async(() => {
     fixture = TestBed.createComponent(<%= classify(name) %>Page);
     component = fixture.componentInstance;
     fixture.detectChanges();
